python_service/requirements.txt
flask==3.0.2
tensorflow==2.15.0
Pillow==10.2.0
numpy==1.26.4
requests==2.31.0
opencv-python-headless==4.9.0.80
